feat: launch app with tab last seen
commit 25c1f1e98c42962ccc23aa08fdf94babba58fc03
Author: Jingchao <alswlx@gmail.com>
Date:   Sat Jan 7 23:25:18 2023 +0800

    feat: restore works

commit 1783b9abaeb1151e2e7a232ea7c1cff9546df7e1
Author: Jingchao <alswlx@gmail.com>
Date:   Sat Jan 7 22:46:02 2023 +0800

    feat: refactory

commit c9a2922a485d0ae82725ffcd5be8974f5d61a0ef
Author: Jingchao <alswlx@gmail.com>
Date:   Sat Jan 7 22:31:43 2023 +0800

    feat: start with tab last seen works

commit 2600ca18a5e6811057cb9873abfe573f4db60e02
Author: Jingchao <alswlx@gmail.com>
Date:   Fri Jan 6 21:43:47 2023 +0800

    feat: setting service

+++ pkg/ui/taskspane/model.go
@@ -36,7 +36,7 @@ func (m Model) Init() tea.Cmd {
 	return nil
 }
 
-func InitModel(tasks []*models.RichTask, width, height int64) Model {
+func InitModel(tasks []*models.RichTask, width, height int) Model {
 	keys := table.DefaultKeyMap()
 	keys.RowDown.SetKeys("j", "down")
 	keys.RowUp.SetKeys("k", "up")
@@ -68,7 +68,7 @@ func InitModel(tasks []*models.RichTask, width, height int64) Model {
 		tableWidth: defaultTableWidth,
 		//props:      app.GetStates(),
 	}
-	m.Resize(int(width), int(height))
+	m.Resize(width, height)
 
 	m.Blur()
